refactor: DB 커넥션 시간을 줄이기 위해 질문 생성 Facade 객체 생성
